requirements.txt
certifi==2020.11.8
chardet==3.0.4
docutils==0.16
idna==2.10
ipaddress==1.0.23
lockfile==0.12.2
marshmallow==3.9.1
pyTenable==1.2.3
python-box==5.2.0
python-daemon==2.2.4
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.0
restfly==1.3.5
semver==2.13.0
six==1.15.0
urllib3==1.26.2
